+++ Common/Gfx/Util/D2DUtil.h
@@ -20,7 +20,7 @@ D2D1_RECT_F ToRectF(const Gdiplus::RectF& rect);
 
 D2D1_POINT_2F AddPoint2F(const D2D1_POINT_2F& point1, const D2D1_POINT_2F& point2);
 
-D2D1_POINT_2F FindEdgePoint(const UINT32 theta, const float left, const float top, const float width, const float height);
+D2D1_POINT_2F FindEdgePoint(const float theta, const float left, const float top, const float width, const float height);
 
 }  // namespace Util
 }  // namespace Gfx
